// panelcfg.gen.ts
// Code generated - EDITING IS FUTILE. DO NOT EDIT.
//
// Generated by:
//     public/app/plugins/gen.go
// Using jennies:
//     TSTypesJenny
//     PluginTsTypesJenny
//
// Run 'make gen-cue' from repository root to regenerate.
//
// Modified to add vertical time axis options

import * as common from '@grafana/schema';

export interface Options extends common.OptionsWithTimezones, common.OptionsWithAnnotations {
  legend: common.VizLegendOptions;
  orientation?: common.VizOrientation;
  timeCompare?: common.TimeCompareOptions;
  tooltip: common.VizTooltipOptions;
  // New options for vertical time axis
  verticalTimeAxis?: boolean;      // Rotacionar eixo de tempo para vertical
  invertTimeDirection?: boolean;   // Inverter direção do tempo (quando vertical)
}

export interface FieldConfig extends common.GraphFieldConfig {}